+++ pkg/client/api/v0043/v0043.go
@@ -25,6 +25,7 @@ type ClientInterface interface {
 	JobInfoInterface
 	NodeInterface
 	PartitionInterface
+	ReconfigureInterface
 	StatsInterface
 }