+++ pkg/client/informer.go
@@ -128,6 +128,8 @@ func (i *informerCache) runInformer(stopCh <-chan struct{}) {
 			list = &types.V0041NodeList{}
 		case types.ObjectTypeV0041PartitionInfo:
 			list = &types.V0041PartitionInfoList{}
+		case types.ObjectTypeV0041Reconfigure:
+			panic("Reconfigure is not supported, this scenario should have been avoided.")
 		case types.ObjectTypeV0041Stats:
 			list = &types.V0041StatsList{}
 
@@ -139,6 +141,8 @@ func (i *informerCache) runInformer(stopCh <-chan struct{}) {
 			list = &types.V0042NodeList{}
 		case types.ObjectTypeV0042PartitionInfo:
 			list = &types.V0042PartitionInfoList{}
+		case types.ObjectTypeV0042Reconfigure:
+			panic("Reconfigure is not supported, this scenario should have been avoided.")
 		case types.ObjectTypeV0042Stats:
 			list = &types.V0042StatsList{}
 
@@ -150,6 +154,8 @@ func (i *informerCache) runInformer(stopCh <-chan struct{}) {
 			list = &types.V0043NodeList{}
 		case types.ObjectTypeV0043PartitionInfo:
 			list = &types.V0043PartitionInfoList{}
+		case types.ObjectTypeV0043Reconfigure:
+			panic("Reconfigure is not supported, this scenario should have been avoided.")
 		case types.ObjectTypeV0043Stats:
 			list = &types.V0043StatsList{}
 
@@ -219,6 +225,8 @@ func (i *informerCache) runGetInformer(stopCh <-chan struct{}) {
 			obj = &types.V0041Node{}
 		case types.ObjectTypeV0041PartitionInfo:
 			obj = &types.V0041PartitionInfo{}
+		// case types.ObjectTypeV0041Reconfigure:
+		// 	obj = &types.V0041Reconfigure{}
 		case types.ObjectTypeV0041Stats:
 			obj = &types.V0041Stats{}
 
@@ -230,6 +238,8 @@ func (i *informerCache) runGetInformer(stopCh <-chan struct{}) {
 			obj = &types.V0042Node{}
 		case types.ObjectTypeV0042PartitionInfo:
 			obj = &types.V0042PartitionInfo{}
+		// case types.ObjectTypeV0042Reconfigure:
+		// 	obj = &types.V0042Reconfigure{}
 		case types.ObjectTypeV0042Stats:
 			obj = &types.V0042Stats{}
 
@@ -241,6 +251,8 @@ func (i *informerCache) runGetInformer(stopCh <-chan struct{}) {
 			obj = &types.V0043Node{}
 		case types.ObjectTypeV0043PartitionInfo:
 			obj = &types.V0043PartitionInfo{}
+		// case types.ObjectTypeV0043Reconfigure:
+		// 	obj = &types.V0043Reconfigure{}
 		case types.ObjectTypeV0043Stats:
 			obj = &types.V0043Stats{}