+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added v43, v42, v41 Reconfigure object.
+
 ### Fixed
 
 - Fixed informer cache not handling v43, v42 objects, and v41 stats.